+++ backend/db/migrations/000042_add_device_prev_next_thumb.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE devices DROP COLUMN prev_thumb_id;
+ALTER TABLE devices DROP COLUMN next_thumb_id;